fix: improve trusted header authentication
