+++ server/auth/header/header.go
@@ -1,11 +1,15 @@
 package header
 
 import (
+	"context"
+	"crypto/subtle"
 	"fmt"
-	"slices"
 	"strings"
 
 	"google.golang.org/grpc/metadata"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 
 	"github.com/argoproj/argo-workflows/v4/config"
 	"github.com/argoproj/argo-workflows/v4/server/auth/types"
@@ -26,12 +30,62 @@ func (h *header) IsRBACEnabled() bool {
 	return h.config.RBAC.IsEnabled()
 }
 
-func New(cfg config.HeaderConfig, sharedSecret string, trustUnauthenticated bool) Interface {
+func getHeaderSharedSecret(
+	ctx context.Context,
+	secretsIf corev1.SecretInterface,
+	cfg *config.SharedSecretHeader,
+) (string, error) {
+	if cfg == nil {
+		return "", fmt.Errorf("shared secret authentication is not configured")
+	}
+	if cfg.Header == "" {
+		return "", fmt.Errorf("shared secret header is empty")
+	}
+	secretRef := cfg.RequiredValue
+	if secretRef.Name == "" || secretRef.Key == "" {
+		return "", fmt.Errorf("shared secret reference is empty")
+	}
+
+	secret, err := secretsIf.Get(ctx, secretRef.Name, metav1.GetOptions{})
+	if err != nil {
+		return "", fmt.Errorf("failed to get shared secret: %w", err)
+	}
+
+	value, ok := secret.Data[secretRef.Key]
+	if !ok {
+		return "", fmt.Errorf(
+			"key %s missing in secret %s",
+			secretRef.Key,
+			secretRef.Name,
+		)
+	}
+
+	if len(value) == 0 {
+		return "", fmt.Errorf("shared secret value is empty")
+	}
+
+	return string(value), nil
+}
+
+func New(ctx context.Context, cfg config.HeaderConfig, secretsIf corev1.SecretInterface, trustUnauthenticated bool) (Interface, error) {
+	var sharedSecret string
+
+	if !trustUnauthenticated {
+		var err error
+		sharedSecret, err = getHeaderSharedSecret(
+			ctx,
+			secretsIf,
+			cfg.SharedSecret,
+		)
+		if err != nil {
+			return nil, err
+		}
+	}
 	return &header{
 		config:               cfg,
 		sharedSecret:         sharedSecret,
 		trustUnauthenticated: trustUnauthenticated,
-	}
+	}, nil
 }
 
 func resolveClaim(source config.ClaimSource, md metadata.MD) string {
@@ -60,10 +114,8 @@ func resolveGroups(source config.GroupClaimSource, md metadata.MD) []string {
 }
 
 func (h *header) Authorize(md metadata.MD) (*types.Claims, error) {
-	if !h.trustUnauthenticated {
-		if err := h.authenticateProxy(md); err != nil {
-			return nil, err
-		}
+	if err := h.authenticateProxy(md); err != nil {
+		return nil, err
 	}
 	claims := &types.Claims{}
 
@@ -82,18 +134,27 @@ func (h *header) Authorize(md metadata.MD) (*types.Claims, error) {
 }
 
 func (h *header) authenticateProxy(md metadata.MD) error {
+	if h.trustUnauthenticated {
+		return nil
+	}
+
 	if h.config.SharedSecret == nil {
 		return fmt.Errorf("shared secret authentication is not configured")
 	}
 
-	values := md.Get(strings.ToLower(h.config.SharedSecret.Header))
+	values := md.Get(h.config.SharedSecret.Header)
 	if len(values) == 0 {
 		return fmt.Errorf("trusted proxy authentication header is missing")
 	}
 
-	if slices.Contains(values, h.sharedSecret) {
-		return nil
+	providedSecret := strings.Join(values, ",")
+
+	if subtle.ConstantTimeCompare(
+		[]byte(providedSecret),
+		[]byte(h.sharedSecret),
+	) != 1 {
+		return fmt.Errorf("trusted proxy authentication failed")
 	}
 
-	return fmt.Errorf("trusted proxy authentication failed")
+	return nil
 }
