+++ server/auth/header/header_test.go
@@ -1,11 +1,15 @@
 package header
 
 import (
+	"context"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc/metadata"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes/fake"
 
 	"github.com/argoproj/argo-workflows/v4/config"
 )
@@ -17,7 +21,6 @@ func TestAuthorize(t *testing.T) {
 		cfg config.HeaderConfig
 		md  metadata.MD
 
-		sharedSecret         string
 		trustUnauthenticated bool
 
 		issuer  string
@@ -153,6 +156,12 @@ func TestAuthorize(t *testing.T) {
 			cfg: config.HeaderConfig{
 				SharedSecret: &config.SharedSecretHeader{
 					Header: "X-Proxy-Auth",
+					RequiredValue: corev1.SecretKeySelector{
+						LocalObjectReference: corev1.LocalObjectReference{
+							Name: "argo-header-auth",
+						},
+						Key: "proxy-secret",
+					},
 				},
 				Subject: config.ClaimSource{
 					Header: "X-Forwarded-User",
@@ -164,15 +173,39 @@ func TestAuthorize(t *testing.T) {
 				"x-forwarded-user", "pradeep",
 			),
 
-			sharedSecret: "secret",
-
 			subject: "pradeep",
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			h := New(tt.cfg, tt.sharedSecret, tt.trustUnauthenticated)
+			clientset := fake.NewSimpleClientset()
+
+			if tt.cfg.SharedSecret != nil {
+				_, err := clientset.CoreV1().
+					Secrets("default").
+					Create(
+						context.Background(),
+						&corev1.Secret{
+							ObjectMeta: metav1.ObjectMeta{
+								Name: "argo-header-auth",
+							},
+							Data: map[string][]byte{
+								"proxy-secret": []byte("secret"),
+							},
+						},
+						metav1.CreateOptions{},
+					)
+				require.NoError(t, err)
+			}
+
+			h, err := New(
+				context.Background(),
+				tt.cfg,
+				clientset.CoreV1().Secrets("default"),
+				tt.trustUnauthenticated,
+			)
+			require.NoError(t, err)
 
 			claims, err := h.Authorize(tt.md)
 
@@ -214,7 +247,15 @@ func TestIsRBACEnabled(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			h := New(tt.cfg, "", true)
+			clientset := fake.NewSimpleClientset()
+
+			h, err := New(
+				context.Background(),
+				tt.cfg,
+				clientset.CoreV1().Secrets("default"),
+				true,
+			)
+			require.NoError(t, err)
 
 			assert.Equal(t, tt.want, h.IsRBACEnabled())
 		})
@@ -228,7 +269,15 @@ func TestAuthorizeMissingSubject(t *testing.T) {
 		},
 	}
 
-	h := New(cfg, "", true)
+	clientset := fake.NewSimpleClientset()
+
+	h, err := New(
+		context.Background(),
+		cfg,
+		clientset.CoreV1().Secrets("default"),
+		true,
+	)
+	require.NoError(t, err)
 
 	claims, err := h.Authorize(metadata.MD{})
 
@@ -240,7 +289,6 @@ func TestAuthenticateProxy(t *testing.T) {
 	tests := []struct {
 		name          string
 		cfg           config.HeaderConfig
-		sharedSecret  string
 		md            metadata.MD
 		expectedError string
 	}{
@@ -249,9 +297,14 @@ func TestAuthenticateProxy(t *testing.T) {
 			cfg: config.HeaderConfig{
 				SharedSecret: &config.SharedSecretHeader{
 					Header: "X-Proxy-Auth",
+					RequiredValue: corev1.SecretKeySelector{
+						LocalObjectReference: corev1.LocalObjectReference{
+							Name: "argo-header-auth",
+						},
+						Key: "proxy-secret",
+					},
 				},
 			},
-			sharedSecret: "secret",
 			md: metadata.Pairs(
 				"x-proxy-auth", "secret",
 			),
@@ -261,9 +314,14 @@ func TestAuthenticateProxy(t *testing.T) {
 			cfg: config.HeaderConfig{
 				SharedSecret: &config.SharedSecretHeader{
 					Header: "X-Proxy-Auth",
+					RequiredValue: corev1.SecretKeySelector{
+						LocalObjectReference: corev1.LocalObjectReference{
+							Name: "argo-header-auth",
+						},
+						Key: "proxy-secret",
+					},
 				},
 			},
-			sharedSecret: "secret",
 			md: metadata.Pairs(
 				"x-proxy-auth", "wrong-secret",
 			),
@@ -274,39 +332,67 @@ func TestAuthenticateProxy(t *testing.T) {
 			cfg: config.HeaderConfig{
 				SharedSecret: &config.SharedSecretHeader{
 					Header: "X-Proxy-Auth",
+					RequiredValue: corev1.SecretKeySelector{
+						LocalObjectReference: corev1.LocalObjectReference{
+							Name: "argo-header-auth",
+						},
+						Key: "proxy-secret",
+					},
 				},
 			},
-			sharedSecret:  "secret",
 			md:            metadata.MD{},
 			expectedError: "trusted proxy authentication header is missing",
 		},
 		{
-			name:          "shared secret not configured",
-			cfg:           config.HeaderConfig{},
-			sharedSecret:  "secret",
-			md:            metadata.Pairs("x-proxy-auth", "secret"),
-			expectedError: "shared secret authentication is not configured",
-		},
-		{
-			name: "one of multiple authentication headers matches",
+			name: "multiple authentication headers",
 			cfg: config.HeaderConfig{
 				SharedSecret: &config.SharedSecretHeader{
 					Header: "X-Proxy-Auth",
+					RequiredValue: corev1.SecretKeySelector{
+						LocalObjectReference: corev1.LocalObjectReference{
+							Name: "argo-header-auth",
+						},
+						Key: "proxy-secret",
+					},
 				},
 			},
-			sharedSecret: "secret",
 			md: metadata.Pairs(
 				"x-proxy-auth", "wrong-secret",
 				"x-proxy-auth", "secret",
 			),
+			expectedError: "trusted proxy authentication failed",
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			h := New(tt.cfg, tt.sharedSecret, false)
+			clientset := fake.NewSimpleClientset()
 
-			err := h.(*header).authenticateProxy(tt.md)
+			_, err := clientset.CoreV1().
+				Secrets("default").
+				Create(
+					context.Background(),
+					&corev1.Secret{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "argo-header-auth",
+						},
+						Data: map[string][]byte{
+							"proxy-secret": []byte("secret"),
+						},
+					},
+					metav1.CreateOptions{},
+				)
+			require.NoError(t, err)
+
+			h, err := New(
+				context.Background(),
+				tt.cfg,
+				clientset.CoreV1().Secrets("default"),
+				false,
+			)
+			require.NoError(t, err)
+
+			err = h.(*header).authenticateProxy(tt.md)
 
 			if tt.expectedError != "" {
 				require.EqualError(t, err, tt.expectedError)
@@ -321,6 +407,12 @@ func TestAuthorizeProxyAuthentication(t *testing.T) {
 	cfg := config.HeaderConfig{
 		SharedSecret: &config.SharedSecretHeader{
 			Header: "X-Proxy-Auth",
+			RequiredValue: corev1.SecretKeySelector{
+				LocalObjectReference: corev1.LocalObjectReference{
+					Name: "argo-header-auth",
+				},
+				Key: "proxy-secret",
+			},
 		},
 		Subject: config.ClaimSource{
 			Header: "X-Forwarded-User",
@@ -328,7 +420,31 @@ func TestAuthorizeProxyAuthentication(t *testing.T) {
 	}
 
 	t.Run("secure mode requires valid proxy secret", func(t *testing.T) {
-		h := New(cfg, "secret", false)
+		clientset := fake.NewSimpleClientset()
+
+		_, err := clientset.CoreV1().
+			Secrets("default").
+			Create(
+				context.Background(),
+				&corev1.Secret{
+					ObjectMeta: metav1.ObjectMeta{
+						Name: "argo-header-auth",
+					},
+					Data: map[string][]byte{
+						"proxy-secret": []byte("secret"),
+					},
+				},
+				metav1.CreateOptions{},
+			)
+		require.NoError(t, err)
+
+		h, err := New(
+			context.Background(),
+			cfg,
+			clientset.CoreV1().Secrets("default"),
+			false,
+		)
+		require.NoError(t, err)
 
 		claims, err := h.Authorize(metadata.Pairs(
 			"x-proxy-auth", "secret",
@@ -340,7 +456,31 @@ func TestAuthorizeProxyAuthentication(t *testing.T) {
 	})
 
 	t.Run("secure mode rejects invalid proxy secret", func(t *testing.T) {
-		h := New(cfg, "secret", false)
+		clientset := fake.NewSimpleClientset()
+
+		_, err := clientset.CoreV1().
+			Secrets("default").
+			Create(
+				context.Background(),
+				&corev1.Secret{
+					ObjectMeta: metav1.ObjectMeta{
+						Name: "argo-header-auth",
+					},
+					Data: map[string][]byte{
+						"proxy-secret": []byte("secret"),
+					},
+				},
+				metav1.CreateOptions{},
+			)
+		require.NoError(t, err)
+
+		h, err := New(
+			context.Background(),
+			cfg,
+			clientset.CoreV1().Secrets("default"),
+			false,
+		)
+		require.NoError(t, err)
 
 		claims, err := h.Authorize(metadata.Pairs(
 			"x-proxy-auth", "wrong-secret",
@@ -352,7 +492,31 @@ func TestAuthorizeProxyAuthentication(t *testing.T) {
 	})
 
 	t.Run("secure mode rejects missing proxy secret", func(t *testing.T) {
-		h := New(cfg, "secret", false)
+		clientset := fake.NewSimpleClientset()
+
+		_, err := clientset.CoreV1().
+			Secrets("default").
+			Create(
+				context.Background(),
+				&corev1.Secret{
+					ObjectMeta: metav1.ObjectMeta{
+						Name: "argo-header-auth",
+					},
+					Data: map[string][]byte{
+						"proxy-secret": []byte("secret"),
+					},
+				},
+				metav1.CreateOptions{},
+			)
+		require.NoError(t, err)
+
+		h, err := New(
+			context.Background(),
+			cfg,
+			clientset.CoreV1().Secrets("default"),
+			false,
+		)
+		require.NoError(t, err)
 
 		claims, err := h.Authorize(metadata.Pairs(
 			"x-forwarded-user", "pradeep",
@@ -363,7 +527,15 @@ func TestAuthorizeProxyAuthentication(t *testing.T) {
 	})
 
 	t.Run("insecure mode skips proxy authentication", func(t *testing.T) {
-		h := New(cfg, "", true)
+		clientset := fake.NewSimpleClientset()
+
+		h, err := New(
+			context.Background(),
+			cfg,
+			clientset.CoreV1().Secrets("default"),
+			true,
+		)
+		require.NoError(t, err)
 
 		claims, err := h.Authorize(metadata.Pairs(
 			"x-forwarded-user", "pradeep",
@@ -372,4 +544,142 @@ func TestAuthorizeProxyAuthentication(t *testing.T) {
 		require.NoError(t, err)
 		assert.Equal(t, "pradeep", claims.Subject)
 	})
+}
+
+func TestGetHeaderSharedSecret(t *testing.T) {
+	tests := []struct {
+		name          string
+		cfg           *config.SharedSecretHeader
+		secret        *corev1.Secret
+		expected      string
+		expectedError string
+	}{
+		{
+			name: "valid secret",
+			cfg: &config.SharedSecretHeader{
+				Header: "X-Proxy-Auth",
+				RequiredValue: corev1.SecretKeySelector{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: "argo-header-auth",
+					},
+					Key: "proxy-secret",
+				},
+			},
+			secret: &corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "argo-header-auth",
+				},
+				Data: map[string][]byte{
+					"proxy-secret": []byte("secret"),
+				},
+			},
+			expected: "secret",
+		},
+		{
+			name:          "shared secret not configured",
+			cfg:           nil,
+			expectedError: "shared secret authentication is not configured",
+		},
+		{
+			name: "secret reference is empty",
+			cfg: &config.SharedSecretHeader{
+				Header: "X-Proxy-Auth",
+			},
+			expectedError: "shared secret reference is empty",
+		},
+		{
+			name: "secret does not exist",
+			cfg: &config.SharedSecretHeader{
+				Header: "X-Proxy-Auth",
+				RequiredValue: corev1.SecretKeySelector{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: "missing-secret",
+					},
+					Key: "proxy-secret",
+				},
+			},
+			expectedError: `failed to get shared secret: secrets "missing-secret" not found`,
+		},
+		{
+			name: "shared secret header is empty",
+			cfg: &config.SharedSecretHeader{
+				RequiredValue: corev1.SecretKeySelector{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: "argo-header-auth",
+					},
+					Key: "proxy-secret",
+				},
+			},
+			expectedError: "shared secret header is empty",
+		},
+		{
+			name: "secret value is empty",
+			cfg: &config.SharedSecretHeader{
+				Header: "X-Proxy-Auth",
+				RequiredValue: corev1.SecretKeySelector{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: "argo-header-auth",
+					},
+					Key: "proxy-secret",
+				},
+			},
+			secret: &corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "argo-header-auth",
+				},
+				Data: map[string][]byte{
+					"proxy-secret": {},
+				},
+			},
+			expectedError: "shared secret value is empty",
+		},
+		{
+			name: "secret key is missing",
+			cfg: &config.SharedSecretHeader{
+				Header: "X-Proxy-Auth",
+				RequiredValue: corev1.SecretKeySelector{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: "argo-header-auth",
+					},
+					Key: "proxy-secret",
+				},
+			},
+			secret: &corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "argo-header-auth",
+				},
+				Data: map[string][]byte{
+					"other-key": []byte("secret"),
+				},
+			},
+			expectedError: "key proxy-secret missing in secret argo-header-auth",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			clientset := fake.NewSimpleClientset()
+
+			if tt.secret != nil {
+				_, err := clientset.CoreV1().
+					Secrets("default").
+					Create(context.Background(), tt.secret, metav1.CreateOptions{})
+				require.NoError(t, err)
+			}
+
+			got, err := getHeaderSharedSecret(
+				context.Background(),
+				clientset.CoreV1().Secrets("default"),
+				tt.cfg,
+			)
+
+			if tt.expectedError != "" {
+				require.EqualError(t, err, tt.expectedError)
+				return
+			}
+
+			require.NoError(t, err)
+			require.Equal(t, tt.expected, got)
+		})
+	}
 }
